Remove Duplicate Bean
Both the autoconfigure module and spring-security-oauth2 produce the
same bean with the same configuration for a session-scoped instance of
OAuth2ClientContext.

Insofar as some will use spring-security-oauth2 without the
autoconfigure module, this commit removes the duplicate bean from
autoconfigure.

Note here that the auto configuration class that generates this bean
already depends on the configuration bean that generates the original
so the final configuration is identical.

Issue: gh-39