Revert of gh-49
Since there is always at least a default UserDetailsService bean
available, its presence in the application context is not
sufficient evidence that the user wants to reconcile incoming JWTs
with a UserDetailService.

At present, there isn't a way to differentiate between a user-provided
UserDetailsService and the default one that is auto-configured.

If a user indeed wants to tie a UserDetailsService to incoming JWTs,
then it's a relatively simple matter of exposing a
JwtTokenStore in a similar fashion to what
ResourceServerTokenServicesConfiguration does.

Fixes: gh-66